Enhance DTOs with EIS mappings and update data types
- Added XML documentation summaries for various DTOs to clarify their purpose and mapping to EIS entities.
- Changed data types for several properties in DTOs from `int` to `long` or `decimal` to improve precision and accommodate larger values.
- Updated `PagedRequest` and `PagedResponse` classes to standardize pagination properties.
- Enhanced `SubmitSalesTransactionResponseData` to include new fields for validation and error handling.
- Refactored inventory-related DTOs to improve clarity and consistency in property naming and types.

+++ src/PointOfSale.App/Services/PosConfigurationService.cs
@@ -390,9 +390,9 @@ public sealed record PosRuntimeContext(
         && !string.IsNullOrWhiteSpace(SiteId)
         && !string.IsNullOrWhiteSpace(SellerTin);
 
-    public int GlobalConfigVersion => Global?.VersionNo > 0 ? Global.VersionNo : 1;
-    public int TerminalConfigVersion => Terminal?.VersionNo > 0 ? Terminal.VersionNo : 1;
-    public int TaxpayerConfigVersion => Taxpayer?.VersionNo > 0 ? Taxpayer.VersionNo : 1;
+    public int GlobalConfigVersion => Global?.VersionNo > 0 ? (int)Global.VersionNo : 1;
+    public int TerminalConfigVersion => Terminal?.VersionNo > 0 ? (int)Terminal.VersionNo : 1;
+    public int TaxpayerConfigVersion => Taxpayer?.VersionNo > 0 ? (int)Taxpayer.VersionNo : 1;
 
     /// <summary>MRA taxRateId for the standard VAT tier — prefers activated config ids (typically <c>A</c>).</summary>
     public string StandardVatTaxRateId =>
